Replace usages of C math functions
Replacing the C standard headers with the C++ standard headers causes various `error: call to 'floor' promotes float to double [performance-type-promotion-in-math-fn,-warnings-as-errors]`, which are fixed by using the C++ std math functions or our own math functions instead of the C math functions.

- Use `absolute` instead of `abs` and `fabs`.
- Use `std::floor` instead of `floor` and `floorf`.
- Use `std::ceil` instead of `ceil`, `ceilf` and `round_ceil`.
- Use `std::round` instead of `round` and `roundf`.
- Use `std::sin` instead of `sin` and `sinf`.
- Use `std::asin` instead of `asin` and `asinf`.
- Use `std::cos` instead of `cos` and `cosf`.
- Use `std::acos` instead of `acos` and `acosf`.
- Use `std::tan` instead of `tan` and `tanf`.
- Use `std::atan` instead of `atan` and `atanf`.
- Use `std::pow` instead of `pow` and `powf`.
- Use `std::log` instead of `log` and `logf`.
- Use `std::log2` instead of `log2` and `log2f`.
- Use `std::log10` instead of `log10` and `log10f`.
- Use `std::pow` instead of `pow` and `powf`.
- Use `std::sqrt` instead of `sqrt` and `sqrtf`.
- Use `std::fmod` instead of `fmod` and `fmodf`.
- Use `direction(Angle)` instead of `vec2(std::cos(Angle), std::sin(Angle))`.
- Use `length(vec2(x, y))` instead of `std::sqrt(x * x + y * y)`.
- Remove unused `NormalizeAngular` and `AngularDistance` functions.

+++ src/base/vmath.h
@@ -87,8 +87,8 @@ template<typename T>
 constexpr inline vector2_base<T> rotate(const vector2_base<T> &a, float angle)
 {
 	angle = angle * pi / 180.0f;
-	float s = sinf(angle);
-	float c = cosf(angle);
+	float s = std::sin(angle);
+	float c = std::cos(angle);
 	return vector2_base<T>((T)(c * a.x - s * a.y), (T)(s * a.x + c * a.y));
 }
 
@@ -106,7 +106,7 @@ constexpr inline T dot(const vector2_base<T> a, const vector2_base<T> &b)
 
 inline float length(const vector2_base<float> &a)
 {
-	return sqrtf(dot(a, a));
+	return std::sqrt(dot(a, a));
 }
 
 constexpr inline float angle(const vector2_base<float> &a)
@@ -115,7 +115,7 @@ constexpr inline float angle(const vector2_base<float> &a)
 		return 0.0f;
 	else if(a.x == 0)
 		return a.y < 0 ? -pi / 2 : pi / 2;
-	float result = atanf(a.y / a.x);
+	float result = std::atan(a.y / a.x);
 	if(a.x < 0)
 		result = result + pi;
 	return result;
@@ -140,7 +140,7 @@ inline vector2_base<float> normalize(const vector2_base<float> &v)
 
 inline vector2_base<float> direction(float angle)
 {
-	return vector2_base<float>(cosf(angle), sinf(angle));
+	return vector2_base<float>(std::cos(angle), std::sin(angle));
 }
 
 typedef vector2_base<float> vec2;
@@ -271,7 +271,7 @@ constexpr inline vector3_base<T> cross(const vector3_base<T> &a, const vector3_b
 //
 inline float length(const vector3_base<float> &a)
 {
-	return sqrtf(dot(a, a));
+	return std::sqrt(dot(a, a));
 }
 
 inline vector3_base<float> normalize(const vector3_base<float> &v)
